Transaction engine redesigned.
A lot of other cosmetic changes carried out as well.  Wrappers introduced
for a lot of app layer functions.

Failing dce unittests disbled.  Will be reintroduced in the updated dce
engine.

Cross transaction matching taken care of.  FPs emanating from these
matches have now disappeared.  Double inspection of transactions taken
care of as well.

+++ src/app-layer-htp.h
@@ -215,8 +215,8 @@ typedef struct HtpState_ {
                                  each connection */
     Flow *f;                /**< Needed to retrieve the original flow when usin HTPLib callbacks */
     uint16_t flags;
-    uint16_t transaction_cnt;
-    uint16_t transaction_done;
+    uint64_t transaction_cnt;
+    uint64_t transaction_done;
     uint16_t store_tx_id;
     FileContainer *files_ts;
     FileContainer *files_tc;
